Make ReplaceLValue methods available outside CheckBoundsDeclarations (#1103)
* Move IgnoreRedundantCasts from BoundsUtil into ExprUtil

* Move getReferentSizeInChars from BoundsUtil to ExprUtil

* Move ConvertToSignedPointerWidth from BoundsUtil to ExprUtil

* Move BoundsUtil class to BoundsUtil.cpp with IsStandardForm method

* Fix typo

* Move CreateBoundsUnknown into BoundsUtil

* Move FindLValue into ExprUtil

* Add ReplaceLValue as a private method of BoundsUtil

* Move ReplaceLValueInBounds into BoundsUtil
