+++ clang/lib/Sema/CMakeLists.txt
@@ -24,6 +24,7 @@ add_clang_library(clangSema
   AnalysisBasedWarnings.cpp
   AvailableFactsAnalysis.cpp
   BoundsAnalysis.cpp
+  BoundsUtils.cpp
   BoundsWideningAnalysis.cpp
   CheckedCAlias.cpp
   CheckedCAnalysesPrepass.cpp
